requirements.txt
openai>=1.93.0
tenacity>=9.1.2
python-dotenv>=1.1.1
aiohttp>=3.12.13
fastapi>=0.115.14
pydantic-settings>=2.10.1